# lib/api-zod/src/generated/types/product.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Tanim Agricultural Supply Chain API
 * OpenAPI spec version: 0.1.0
 */

export interface Product {
  id: number;
  name: string;
  description?: string | null;
  price: number;
  stock_quantity: number;
  unit: string;
  image_url?: string | null;
  category_id?: number | null;
  category_name?: string | null;
  farmer_id: number;
  farmer_name: string;
  farm_name?: string | null;
  average_rating?: number | null;
  review_count: number;
  low_stock_threshold: number;
  is_active: boolean;
  created_at: Date;
}